// astro-grader/src/lib/types/Nights.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AstroSession } from "./AstroSession";
import type { File } from "./File";

export type Nights = { "PreviewNights": { [key in string]: Array<File> } } | { "GroupedNights": Array<AstroSession> };